Update orbital view
After a lot of trial and error the orbital view is almost complete, the
main issue with correctly determining the position of the vessel was the
wrong value for the longitude of the ascending node.

Using a value of zero ended up working correctly, likely because the
ellipse isn't drawn with this value being considered. This may have to
be fixed in a later revision - it looks right *now*, but may end up not
working correctly in other scenarios.